Move organization read repository to neo4j module
