+++ packages/server/customer-os-neo4j-repository/repository/repositories.go
@@ -22,6 +22,7 @@ type Repositories struct {
 	MasterPlanReadRepository          MasterPlanReadRepository
 	MasterPlanWriteRepository         MasterPlanWriteRepository
 	OpportunityReadRepository         OpportunityReadRepository
+	OrganizationReadRepository        OrganizationReadRepository
 	PhoneNumberReadRepository         PhoneNumberReadRepository
 	PlayerWriteRepository             PlayerWriteRepository
 	ServiceLineItemReadRepository     ServiceLineItemReadRepository
@@ -52,6 +53,7 @@ func InitNeo4jRepositories(driver *neo4j.DriverWithContext, neo4jDatabase string
 		MasterPlanReadRepository:          NewMasterPlanReadRepository(driver, neo4jDatabase),
 		MasterPlanWriteRepository:         NewMasterPlanWriteRepository(driver, neo4jDatabase),
 		OpportunityReadRepository:         NewOpportunityReadRepository(driver, neo4jDatabase),
+		OrganizationReadRepository:        NewOrganizationReadRepository(driver, neo4jDatabase),
 		PhoneNumberReadRepository:         NewPhoneNumberReadRepository(driver, neo4jDatabase),
 		PlayerWriteRepository:             NewPlayerWriteRepository(driver, neo4jDatabase),
 		ServiceLineItemReadRepository:     NewServiceLineItemReadRepository(driver, neo4jDatabase),
